move query and json_encoder to bokeh.core
