+++ bokeh/core/query.py
