+++ bokeh/document.py
@@ -13,13 +13,13 @@ import uuid
 
 from six import string_types
 
+from .core.json_encoder import serialize_json
+from .core.query import find
 from .model import Model
-from .query import find
 from .validation import check_integrity
 from .util.callback_manager import _check_callback
 from .util.deprecate import deprecated
 from .util.version import __version__
-from ._json_encoder import serialize_json
 from .themes import default as default_theme
 from .themes import Theme
