+++ bokeh/models/plots.py
@@ -6,12 +6,12 @@ from __future__ import absolute_import
 from six import string_types
 import warnings
 
+from ..core.query import find
 from ..enums import Location
 from ..mixins import LineProps, TextProps, FillProps
 from ..model import Model
 from ..properties import (Bool, Int, String, Enum, Auto, Instance, Either,
     List, Dict, Include, Override)
-from ..query import find
 from ..util.string import nice_join
 from ..validation.warnings import (MISSING_RENDERERS, NO_GLYPH_RENDERERS,
     EMPTY_LAYOUT, MALFORMED_CATEGORY_LABEL)
